+++ Libraries/SideScroll.Serialize/Atlas/TypeRepos/TypeRepoTimeOnly.cs
@@ -2,7 +2,7 @@ using SideScroll.Serialize.Atlas.Schema;
 
 namespace SideScroll.Serialize.Atlas.TypeRepos;
 
-public class TypeRepoDecimal(Serializer serializer, TypeSchema typeSchema) : TypeRepo(serializer, typeSchema)
+public class TypeRepoTimeOnly(Serializer serializer, TypeSchema typeSchema) : TypeRepo(serializer, typeSchema)
 {
 	public class Creator : IRepoCreator
 	{
@@ -10,7 +10,7 @@ public class TypeRepoDecimal(Serializer serializer, TypeSchema typeSchema) : Typ
 		{
 			if (CanAssign(typeSchema.Type))
 			{
-				return new TypeRepoDecimal(serializer, typeSchema);
+				return new TypeRepoTimeOnly(serializer, typeSchema);
 			}
 			return null;
 		}
@@ -18,12 +18,12 @@ public class TypeRepoDecimal(Serializer serializer, TypeSchema typeSchema) : Typ
 
 	public static bool CanAssign(Type? type)
 	{
-		return type == typeof(decimal);
+		return type == typeof(TimeOnly);
 	}
 
 	public override void SaveObject(BinaryWriter writer, object obj)
 	{
-		writer.Write((decimal)obj);
+		writer.Write(((TimeOnly)obj).Ticks);
 	}
 
 	protected override object? CreateObject(int objectIndex)
@@ -31,16 +31,15 @@ public class TypeRepoDecimal(Serializer serializer, TypeSchema typeSchema) : Typ
 		long position = Reader!.BaseStream.Position;
 		Reader.BaseStream.Position = ObjectOffsets![objectIndex];
 
-		object obj = Reader.ReadDecimal();
+		object obj = new TimeOnly(Reader.ReadInt64());
 		Reader.BaseStream.Position = position;
 
 		ObjectsLoaded[objectIndex] = obj; // must assign before loading any more refs
 		return obj;
 	}
 
+	// not called, it's a struct and a value
 	public override void Clone(object source, object dest)
 	{
-		// assigning won't do anything since it's not a ref
-		throw new SerializerException("Not cloneable");
 	}
 }